+++ test/unit/integrations/payu_in/payu_in_helper_test.rb
@@ -11,6 +11,7 @@ class PayuInHelperTest < Test::Unit::TestCase
     assert_equal '10.00', @helper.fields['amount']
     assert_equal 'merchant_id', @helper.fields['key']
     assert_equal 'order_id', @helper.fields['txnid']
+    assert_equal 'ActiveMerchant', @helper.fields['udf5']
   end
 
   def test_customer_fields
@@ -74,5 +75,16 @@ class PayuInHelperTest < Test::Unit::TestCase
   def test_phone_replace_non_digits
     @helper.fields['phone'] = '+(999)-99 99999'
     assert_equal '9999999999', @helper.form_fields['phone']
+  end
+
+  def test_override_application_id
+    old_identifier = PayuIn::Helper.application_id
+    new_identifier = 'shopping_cart_software'
+
+    PayuIn::Helper.application_id = new_identifier
+    @helper = PayuIn::Helper.new(1,'johndoe@example.com', :amount => 500, :currency => 'INR')
+    assert_field 'udf5', new_identifier
+  ensure
+    PayuIn::Helper.application_id = old_identifier
   end
 end